Refactored HasId
