+++ AirBnb/src/Entities/Booking.java
@@ -1,7 +1,5 @@
 package Entities;
 
-import Repository.HasId;
-
 import java.util.Date;
 
 public class Booking implements HasId, Payable{
