+++ AirBnb/src/Repository/InMemoryRepo.java
@@ -1,5 +1,7 @@
 package Repository;
 
+import Entities.HasId;
+
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
